+++ data/src/main/java/com/whyranoid/data/di/GroupModule.kt
@@ -5,6 +5,8 @@ import com.whyranoid.data.group.GroupDataSourceImpl
 import com.whyranoid.data.group.GroupRepositoryImpl
 import com.whyranoid.data.groupnotification.GroupNotificationDataSource
 import com.whyranoid.data.groupnotification.GroupNotificationDataSourceImpl
+import com.whyranoid.data.user.UserDataSource
+import com.whyranoid.data.user.UserDataSourceImpl
 import com.whyranoid.domain.repository.GroupRepository
 import dagger.Binds
 import dagger.Module
@@ -17,6 +19,9 @@ abstract class GroupModule {
 
     @Binds
     abstract fun bindGroupDataSource(groupDataSourceImpl: GroupDataSourceImpl): GroupDataSource
+
+    @Binds
+    abstract fun bindUserDataSource(userDataSourceImpl: UserDataSourceImpl): UserDataSource
 
     @Binds
     abstract fun bindGroupNotificationDataSource(groupNotificationDataSourceImpl: GroupNotificationDataSourceImpl): GroupNotificationDataSource
